# src/main/java/com/ingestionservice/api/controller/StoreContentsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.18).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.ingestionservice.api.controller;

import com.ingestionservice.api.models.AcmRequest;
import com.ingestionservice.api.models.AcmResponse;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-04-19T16:57:57.183235500+05:30[Asia/Calcutta]")
@Api(value = "store-contents", description = "the store-contents API")
public interface StoreContentsApi {

    @ApiOperation(value = "Get the files from aws s3 bucket and store into monogodb", nickname = "storedContents", notes = "Get the files from aws s3 bucket and store into monogodb", response = AcmResponse.class, authorizations = {
        @Authorization(value = "oAuth2Client", scopes = { 
            }),
@Authorization(value = "oAuth2Password", scopes = { 
            })    }, tags={ "Ingestion", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Operation completed successfully", response = AcmResponse.class),
        @ApiResponse(code = 400, message = "Bad Request"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Access Denied"),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/store-contents",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<AcmResponse> storedContents(@ApiParam(value = "The Ingestion request body." ,required=true )  @Valid @RequestBody AcmRequest body
);

}